+++ controllers/casosController.js
@@ -1,19 +1,17 @@
 const casosRepository = require("../repositories/casosRepository")
 const validations = require("../utils/errorHandler")
 
-function checkExist(id, res){
+function checkExist(id, res) {
     const caso = casosRepository.getCaseByID(id);
     if (!caso) {
-        return res.status(404).json({
-            message: "Caso não cadastrado no banco de dados!"
-        });
-    }s
+        res.status(404).json({ message: "Caso não cadastrado no banco de dados!" });
+        return null;
+    }
     return caso; 
 }
 
 
 function getCasosController(req, res) {
-
         const casos = casosRepository.getAll()
         res.json(casos)
 }
@@ -27,7 +25,10 @@ function getCaseByIDController(req, res) {
 
 function createCaseController(req,res){
         const data = req.body;
-        validations.validationsCase(data, res);
+        const isValid =  validations.validationsCase(data, res);
+        if (!isValid) {
+        return; 
+        }
 
         const newCase = casosRepository.createCase(data);
         res.status(201).json(newCase);
@@ -36,9 +37,12 @@ function createCaseController(req,res){
 function updateCaseController(req,res){
         const { id } = req.params;
         const data = req.body;
-        checkExist(id);
-        validations.validationsCase(data, res);
-  
+        const caso = checkExist(id, res);
+        if (!caso) return; 
+        const isValid =  validations.validationsCase(data, res);
+        if (!isValid) {
+        return; 
+        }
 
         const updatedCaseController = casosRepository.updateCase(id, data);
         res.status(200).json(updatedCaseController);
@@ -47,8 +51,12 @@ function updateCaseController(req,res){
 function patchCaseController(req,res){
         const { id } = req.params;
         const data = req.body;
-        checkExist(id);
-        validations.validationsCase(data, res);
+        const caso = checkExist(id, res);
+        if (!caso) return; 
+        const isValid =  validations.validationsCase(data, res);
+        if (!isValid) {
+        return; 
+        }
         
  
         const patchedCaseController = casosRepository.patchCase(id, data);
@@ -57,10 +65,11 @@ function patchCaseController(req,res){
 
 function deleteCaseController(req,res){
         const { id } = req.params;
-        checkExist(id);
+        const caso = checkExist(id, res);
+        if (!caso) return; 
 
-        const deletedCase = casosRepository.deleteCase(id);
-        res.status(200).json(deletedCase);
+        casosRepository.deleteCase(id);
+        res.status(204).json(id);
         
 }
